requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
distlib==0.3.1
Django==3.2.3
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
filelock==3.0.12
Pillow==8.2.0
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.6
